refactor(cli): split commands.rs into cli/args.rs and cli/handlers.rs
- cli/args.rs: clap derive structs, enums, From impls
- cli/handlers.rs: per-command helpers with flattened run() (1-level nesting)
- cli/mod.rs: facade re-exporting args and run

Preserves exact CLI JSON contracts, exit codes, and error messages.

QA: cargo test -p biors passes (32 tests)
